load file with edit

+++ app/Http/Controllers/FileController.php
@@ -3,6 +3,8 @@
 namespace App\Http\Controllers;
 
 use App\Http\Requests\StoreFileRequest;
+use App\Http\Requests\UpdateFileRequest;
+use App\Http\Resources\FileResource;
 use App\Models\File;
 use App\Models\TempFile;
 use Illuminate\Http\Request;
@@ -20,23 +22,16 @@ class FileController extends Controller
      */
     public function index()
     {
-        $url = Storage::url('62f6ee8cf286e-1660350092/avatar.jpg');
-        Log::build([
-            'driver' => 'single',
-            'path' => storage_path('logs/custom.log'),
-        ])->info($url);
-
         $files = File::orderBy('id'/*, 'DESC'*/)->get();
 
-        return Inertia::render('FilesList', [ 'files' => $files, 'csrf_token' => csrf_token() ]);
+        return Inertia::render('FilesList', ['files' => FileResource::collection($files), 'csrf_token' => csrf_token(), ]);
     }
-
 
 
     /**
      * Store a newly created resource in storage.
      *
-     * @param  \Illuminate\Http\Request  $request
+     * @param \Illuminate\Http\Request $request
      * @return \Illuminate\Http\Response
      */
     public function store(StoreFileRequest $request)
@@ -45,9 +40,12 @@ class FileController extends Controller
 
         $tempFile = TempFile::where('folder', $validated['new_file'])->first();
 
+        $idPath = $tempFile->folder . '/' . $tempFile->filename;
+        $fromPath = 'tmp/' . $idPath;
+        $toPath = 'files/' . $idPath;
 
-        Storage::move('tmp/' . $tempFile->folder . '/' . $tempFile->filename, 'files/' . $tempFile->folder . '/' . $tempFile->filename);
-        Storage::deleteDirectory('tmp/' . $tempFile->folder );
+        Storage::move($fromPath, $toPath);
+        Storage::deleteDirectory('tmp/' . $tempFile->folder);
 
         Auth::user()->files()->create([
             'folder' => $tempFile->folder,
@@ -63,24 +61,43 @@ class FileController extends Controller
     /**
      * Show the form for editing the specified resource.
      *
-     * @param  int  $id
+     * @param int $id
      * @return \Illuminate\Http\Response
      */
-    public function update(Request $request, File $file)
+    public function update(UpdateFileRequest $request, File $file)
     {
+        $validated = $request->validated();
 
-        $tempFile = TempFile::where('folder', $request->new_file)->first();
-        Storage::move('tmp/' . $tempFile->folder . '/' . $tempFile->filename, 'files/' . $tempFile->folder . '/' . $tempFile->filename);
-        Storage::deleteDirectory('tmp/' . $tempFile->folder );
-        Storage::deleteDirectory('files/' . $file->folder );
+        Log::build([
+            'driver' => 'single',
+            'path' => storage_path('logs/custom.log'),
+        ])->info($validated);
 
-        $file->update([
-            'folder' => $tempFile->folder,
-            'filename' => $tempFile->filename,
-            'title' => $request->title,
-        ]);
+        if ($file->folder === $validated["new_file"]) {
+            $file->update([
+                'title' => $validated["title"],
+            ]);
+        } else {
+            $tempFile = TempFile::where('folder', $validated["new_file"])->first();
 
-        $tempFile->delete();
+
+            $idPath = $tempFile->folder . '/' . $tempFile->filename;
+            $fromPath = 'tmp/' . $idPath;
+            $toPath = 'files/' . $idPath;
+
+            Storage::move($fromPath, $toPath);
+            Storage::deleteDirectory('tmp/' . $tempFile->folder);
+            Storage::deleteDirectory('files/' . $file->folder);
+
+            $file->update([
+                'folder' => $tempFile->folder,
+                'filename' => $tempFile->filename,
+                'title' => $validated["title"],
+            ]);
+
+            $tempFile->delete();
+        }
+
 
         return redirect('/upload');
     }
@@ -88,13 +105,29 @@ class FileController extends Controller
     /**
      * Remove the specified resource from storage.
      *
-     * @param  int  $id
+     * @param int $id
      * @return \Illuminate\Http\Response
      */
     public function destroy(File $file)
     {
-        Storage::deleteDirectory('files/' . $file->folder );
+        Storage::deleteDirectory('files/' . $file->folder);
         $file->delete();
         return redirect('/upload');
+    }
+
+    public function load($serverId)
+    {
+        $user = Auth::user();
+
+        $file = File::where('folder', $serverId)->first();
+
+        if ($user->can('load', $file)) {
+            return Storage::download("files/" . $file->folder . "/" . $file->filename);
+        }
+        return "";
+
+
+
+
     }
 }
